requirements.txt
absl-py==1.2.0
affine==2.3.1
aioflask==0.4.0
altair==4.2.0
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
branca==0.5.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.5
cryptography==37.0.4
cv==1.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
ez-setup==0.9
Fiona==1.8.21
Flask==2.1.2
flatbuffers==2.0.7
folium==0.12.1.post1
fonttools==4.37.1
gast==0.4.0
# GDAL==3.5.1
geopandas==0.10.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
greenletio==0.9.0
grpcio==1.48.1
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter-client==6.1.12
jupyter-console==6.1.0
jupyter-core==4.11.1
jupyterlab-widgets==3.0.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
libpysal==4.6.2
mapclassify==2.4.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
# matplotlib-inline==0.1.6
momepy==0.5.3
munch==2.5.0
munkres==1.1.4
nest-asyncio==1.5.5
# networkx==2.6.3
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
# osmnx==1.1.2
packaging==21.3
# pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
# Pillow==9.2.0
pip==22.2.2
prompt-toolkit==3.0.30
protobuf==3.19.5
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygeos==0.13
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
# pyproj==3.3.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
# rasterio==1.3.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.0.0
# scikit-learn==1.1.2
# scipy==1.9.0
setuptools==65.3.0
Shapely==1.8.4
six==1.16.0
snuggs==1.4.7
soupsieve==2.3.2.post1
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
unicodedata2==14.0.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
win-inet-pton==1.1.0
wrapt==1.14.1
xyzservices==2022.6.0
